refactor: initial refactor of transcript viewer
