+++ src/atoms/side-panel.ts
@@ -4,7 +4,7 @@ import { atom } from 'jotai'
 
 import { ConversationData } from '@/types/conversations'
 
-export const transcriptOpenAtom = atom(false)
+export const sidePanelOpenAtom = atom(false)
 
 export const selectedAudioNoteAtom = atom<Partial<ConversationData> | null>(null)
