+++ src/features/app/hooks/use-auth-change-handler.ts
