+++ src/features/app/hooks/use-initialize-amplitude-and-sentry.ts
